+++ Library/RaisrHandler.h
@@ -14,7 +14,7 @@ extern "C"
 
     extern RNLERRORTYPE RNLHandler_Init(
         const char *modelPath,
-        unsigned int ratio,
+        float ratio,
         unsigned int bitDepth,
         RangeType    rangeType,
         unsigned int threadCount,
